fix: move authOptions to lib
